% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{arima_all}
\alias{arima_all}
\alias{arima_all-package}
\title{arima_all.}
\usage{
arima_all(x)
}
\arguments{
\item{x}{vector of observations}
}
\description{
arima_all.

This function returns some diagnostics from auto.arima as JSON
serializable list.
}
\examples{
arima_all(WWWusage)
}
\keyword{arima}

