% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{arima_all}
\alias{arima_all}
\title{An arima function}
\usage{
arima_all(x)
}
\arguments{
\item{x}{vector of observations}
}
\description{
This function returns some diagnostics from auto.arima as JSON
serializable list.
}
\examples{
arima_all(WWWusage)
}
\keyword{arima}

